\name{dict-package}
\alias{dict-package}
\alias{dict}
\docType{package}
\title{
\packageTitle{dict}
}
\description{
\packageDescription{dict}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{dict}
\packageIndices{dict}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{dict}

Maintainer: \packageMaintainer{dict}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
