% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{dict}
\alias{dict}
\title{Create word dictionary}
\usage{
dict(x)
}
\arguments{
\item{x}{Input can be a vector of words, a list of 'positive' and 'negative'
word vectors, a data frame with 'word' and 'weight' columns}
}
\value{
A dictionary
}
\description{
Create word dictionary
}
\examples{

pos <- c("like", "love", "amazing", "excellent", "great", "fantastic",
  "incredible", "awesome", "best", "favorite", "fan", "fun", "enjoyed",
  "enjoyed", "good", "solid", "better", "soooo", "happy")
neg <- c("hate", "loathe", "dislike", "awful", "horrible", "worst",
  "miserable", "ruin", "ruining", "destroy", "destroyed", "destroying",
  "pathetic", "hated", "hateful", "unhappy", "horrifies", "horrifying",
  "terrible")

## create dictionary using only positive words
dict(pos)

## create dictionary using only negative words
dict(neg)

## create dictionary using both positive and negative words
d <- dict(list(pos = pos, neg = neg))

## view up to n entries of dictionary
print(d, n = 100)

## example text
txt <- c("love amazing excellent good",
  "hate loathe horrifies unhappy terrible",
  "awesome best hateful hated worst")

## get estimate for each element of txt using pos/neg dictionary
d(txt)

## create tibble
tibble::tibble(
  text = txt,
  sent = d(txt)
)

}
