% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sent.R
\name{dict_apply}
\alias{dict_apply}
\title{Use dictionary}
\usage{
dict_apply(x, dict)
}
\arguments{
\item{x}{Input text vector}

\item{dict}{A word dictionary}
}
\value{
Counts.
}
\description{
Calculates counts based on word dictionaries
}
\keyword{internal}
