% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sent.R
\name{dict_score}
\alias{dict_score}
\alias{dict_score_score}
\title{Use dictionary}
\usage{
dict_score(x, dict)

dict_score_score(x, dict)
}
\arguments{
\item{x}{Input text vector}

\item{dict}{A word dictionary}
}
\value{
A data frame with the number of postively and negatively coded words,
  the overall score (sum of positive and negative), and the total word count.
}
\description{
Calculates counts based on word dictionaries
}
\keyword{internal}
